@@ -1,6 +1,6 @@
 {
-  "name": "angular-advanced-workshop",
-  "version": "13.0.0",
+  "name": "bookmonkey-client-advanced",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -12,49 +12,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.1.1",
-    "@angular/cdk": "13.1.1",
-    "@angular/common": "13.1.1",
-    "@angular/compiler": "13.1.1",
-    "@angular/core": "13.1.1",
-    "@angular/forms": "13.1.1",
-    "@angular/material": "13.1.1",
-    "@angular/platform-browser": "13.1.1",
-    "@angular/platform-browser-dynamic": "13.1.1",
-    "@angular/router": "13.1.1",
-    "core-js": "3.6.5",
-    "rxjs": "7.5.1",
-    "tslib": "2.3.1",
-    "zone.js": "0.11.4"
+    "@angular/animations": "^17.1.2",
+    "@angular/cdk": "^17.1.2",
+    "@angular/common": "^17.1.2",
+    "@angular/compiler": "^17.1.2",
+    "@angular/core": "^17.1.2",
+    "@angular/forms": "^17.1.2",
+    "@angular/material": "^17.1.2",
+    "@angular/platform-browser": "^17.1.2",
+    "@angular/platform-browser-dynamic": "^17.1.2",
+    "@angular/router": "^17.1.2",
+    "@ngrx/effects": "^17.1.0",
+    "@ngrx/entity": "^17.1.0",
+    "@ngrx/router-store": "^17.1.0",
+    "@ngrx/store": "^17.1.0",
+    "@ngrx/store-devtools": "^17.1.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.1.2",
-    "@angular-eslint/builder": "13.0.1",
-    "@angular-eslint/eslint-plugin": "13.0.1",
-    "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/schematics": "13.0.1",
-    "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "13.1.2",
-    "@angular/compiler-cli": "13.1.1",
-    "@angular/language-service": "13.1.1",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "2.0.8",
-    "@types/node": "14.6.3",
-    "@typescript-eslint/eslint-plugin": "5.8.1",
-    "@typescript-eslint/parser": "5.8.1",
-    "eslint": "8.5.0",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-prettier": "4.0.0",
-    "jasmine-core": "3.6.0",
+    "@angular-devkit/build-angular": "^17.1.2",
+    "@angular-eslint/builder": "^17.2.1",
+    "@angular-eslint/eslint-plugin": "^17.2.1",
+    "@angular-eslint/eslint-plugin-template": "^17.2.1",
+    "@angular-eslint/schematics": "^17.2.1",
+    "@angular-eslint/template-parser": "^17.2.1",
+    "@angular/cli": "^17.1.2",
+    "@angular/compiler-cli": "^17.1.2",
+    "@ngneat/spectator": "^16.0.0",
+    "@ngrx/eslint-plugin": "^17.1.0",
+    "@types/jasmine": "~5.1.0",
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
+    "eslint": "^8.53.0",
+    "eslint-config-prettier": "~9.1.0",
+    "eslint-plugin-prettier": "~5.1.3",
+    "jasmine-core": "~5.1.0",
     "jasmine-spec-reporter": "6.0.0",
-    "karma": "6.3.9",
-    "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "3.0.3",
-    "karma-jasmine": "4.0.1",
-    "karma-jasmine-html-reporter": "1.5.4",
-    "prettier": "2.1.1",
-    "protractor": "7.0.0",
-    "ts-node": "9.0.0",
-    "typescript": "~4.5.4"
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "prettier": "~3.2.4",
+    "typescript": "~5.3.3"
   }
 }
