@@ -1,34 +1,63 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
-import { Subscription } from 'rxjs';
-import { switchMap } from 'rxjs/operators';
+import { Component, DestroyRef, Input } from '@angular/core';
+import { RouterLink } from '@angular/router';
+import { EMPTY, Observable } from 'rxjs';
+import { tap } from 'rxjs/operators';
 import { BookApiService } from '../book-api.service';
-import { Book, bookNa } from '../models';
+import { Book } from '../models';
+import { MatButton } from '@angular/material/button';
+import { MatInput, MatLabel } from '@angular/material/input';
+import { MatError, MatFormField } from '@angular/material/form-field';
+import { FormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
+import { AsyncPipe, NgIf } from '@angular/common';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 
 @Component({
   selector: 'ws-book-edit',
   templateUrl: './book-edit.component.html',
-  styleUrls: ['./book-edit.component.scss']
+  styleUrls: ['./book-edit.component.scss'],
+  standalone: true,
+  imports: [NgIf, ReactiveFormsModule, MatFormField, MatInput, MatLabel, MatError, MatButton, RouterLink, AsyncPipe]
 })
-export class BookEditComponent implements OnInit, OnDestroy {
-  sink = new Subscription();
-  book: Book = bookNa();
+export class BookEditComponent {
+  protected book$: Observable<Book> = EMPTY;
+  protected isbnValue = '';
 
-  constructor(private route: ActivatedRoute, private bookService: BookApiService) {}
+  protected form = this.formBuilder.nonNullable.group({
+    title: ['', [Validators.required]],
+    subtitle: [''],
+    author: ['', [Validators.required]],
+    abstract: [''],
+    isbn: [''],
+    cover: ['']
+  });
 
-  ngOnInit() {
-    this.sink.add(
-      this.route.params
-        .pipe(switchMap(params => this.bookService.getByIsbn(params.isbn)))
-        .subscribe(book => (this.book = book))
+  constructor(
+    private readonly formBuilder: FormBuilder,
+    private readonly bookService: BookApiService,
+    private readonly destroyRef: DestroyRef
+  ) {}
+
+  @Input({ required: true })
+  set isbn(isbn: string) {
+    this.book$ = this.bookService.getByIsbn(isbn).pipe(
+      tap(book => {
+        this.form.setValue({
+          title: book.title,
+          subtitle: book.subtitle,
+          author: book.author,
+          abstract: book.abstract,
+          isbn: book.isbn,
+          cover: book.cover
+        });
+      })
     );
-  }
-
-  ngOnDestroy() {
-    this.sink.unsubscribe();
+    this.isbnValue = isbn;
   }
 
   save() {
-    this.sink.add(this.bookService.update(this.book.isbn, this.book).subscribe());
+    this.bookService
+      .update(this.isbnValue, this.form.getRawValue())
+      .pipe(takeUntilDestroyed(this.destroyRef))
+      .subscribe();
   }
 }
