@@ -2,16 +2,20 @@ import { Component } from '@angular/core';
 import { Observable } from 'rxjs';
 import { BookApiService } from '../book-api.service';
 import { Book } from '../models';
+import { BookCardComponent } from '../book-card/book-card.component';
+import { AsyncPipe, NgFor } from '@angular/common';
 
 @Component({
   selector: 'ws-book-list',
   styleUrls: ['./book-list.component.scss'],
-  templateUrl: 'book-list.component.html'
+  templateUrl: 'book-list.component.html',
+  standalone: true,
+  imports: [NgFor, BookCardComponent, AsyncPipe]
 })
 export class BookListComponent {
-  books$: Observable<Book[]>;
+  protected books$: Observable<Book[]>;
 
-  constructor(private bookData: BookApiService) {
-    this.books$ = this.bookData.getAll();
+  constructor(private readonly bookService: BookApiService) {
+    this.books$ = this.bookService.getAll();
   }
 }
