@@ -1,44 +1,47 @@
-import { Component, OnDestroy } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { Router } from '@angular/router';
-import { Subscription } from 'rxjs';
+import { Component, DestroyRef } from '@angular/core';
+import { FormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
+import { Router, RouterLink } from '@angular/router';
 import { tap } from 'rxjs/operators';
 import { BookApiService } from '../book-api.service';
 import { bookNa } from '../models';
+import { MatButton } from '@angular/material/button';
+import { NgIf } from '@angular/common';
+import { MatInput, MatLabel } from '@angular/material/input';
+import { MatError, MatFormField } from '@angular/material/form-field';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 
 @Component({
   selector: 'ws-book-new',
   styleUrls: ['./book-new.component.scss'],
-  templateUrl: './book-new.component.html'
+  templateUrl: './book-new.component.html',
+  standalone: true,
+  imports: [ReactiveFormsModule, MatFormField, MatInput, NgIf, MatError, MatButton, RouterLink, MatLabel]
 })
-export class BookNewComponent implements OnDestroy {
-  sink = new Subscription();
-  form: FormGroup;
+export class BookNewComponent {
+  protected form = this.formBuilder.nonNullable.group({
+    title: ['', [Validators.required]],
+    subtitle: [''],
+    author: ['', [Validators.required]],
+    abstract: [''],
+    isbn: ['', [Validators.required, Validators.minLength(3)]],
+    cover: ['']
+  });
 
-  constructor(private router: Router, private fb: FormBuilder, private bookService: BookApiService) {
-    this.form = this.buildForm();
-  }
-
-  ngOnDestroy() {
-    this.sink.unsubscribe();
-  }
+  constructor(
+    private readonly formBuilder: FormBuilder,
+    private readonly router: Router,
+    private readonly bookService: BookApiService,
+    private readonly destroyRef: DestroyRef
+  ) {}
 
   create() {
-    const book = { ...bookNa(), ...this.form.value };
-    this.sink.add(
-      this.bookService
-        .create(book)
-        .pipe(tap(() => this.router.navigateByUrl('/')))
-        .subscribe()
-    );
-  }
-
-  private buildForm(): FormGroup {
-    return this.fb.group({
-      isbn: ['', [Validators.required, Validators.minLength(3)]],
-      title: ['', Validators.required],
-      author: ['', Validators.required],
-      cover: ['']
-    });
+    const book = { ...bookNa(), ...this.form.getRawValue() };
+    this.bookService
+      .create(book)
+      .pipe(
+        takeUntilDestroyed(this.destroyRef),
+        tap(() => this.router.navigateByUrl('/'))
+      )
+      .subscribe();
   }
 }
