@@ -1,7 +1,10 @@
 import { Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
 
 @Component({
   selector: 'ws-book',
-  templateUrl: './book.component.html'
+  templateUrl: './book.component.html',
+  standalone: true,
+  imports: [RouterOutlet]
 })
 export class BookComponent {}
