@@ -11,6 +11,7 @@ export const bookNa: () => Book = () => {
     title: na,
     subtitle: na,
     numPages: 0,
-    publisher: { name: na, url: na }
+    publisher: na,
+    price: 0
   };
 };
